requirements.txt
openai>=1.10.0
streamlit>=1.26.0
faiss-cpu>=1.7.4
spacy>=3.6.0
networkx>=3.1
pdfplumber>=0.10.0
tqdm>=4.66.1
python-dotenv>=1.0.0
langdetect>=1.0.9
PyPDF2>=3.0.0
matplotlib>=3.5.0
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
PyArabic>=0.6.15
transformers>=4.28.0
torch>=2.0.0